+++ old-ui/app/components/notice.js
@@ -14,7 +14,7 @@ function Notice () {
 
 Notice.prototype.render = function () {
   const { notice, onConfirm } = this.props
-  const { title, date, body } = notice
+  const { title, body } = notice
   const state = this.state || { disclaimerDisabled: false }
   const disabled = state.disclaimerDisabled
 
@@ -22,35 +22,9 @@ Notice.prototype.render = function () {
     h('.flex-column.flex-center.flex-grow', {
       style: {
         width: '100%',
+        fontFamily: 'Nunito SemiBold',
       },
     }, [
-      h('h3.flex-center.text-transform-uppercase.terms-header', {
-        style: {
-          background: '#EBEBEB',
-          color: '#AEAEAE',
-          width: '100%',
-          fontSize: '20px',
-          textAlign: 'center',
-          padding: 6,
-        },
-      }, [
-        title,
-      ]),
-
-      h('h5.flex-center.text-transform-uppercase.terms-header', {
-        style: {
-          background: '#EBEBEB',
-          color: '#AEAEAE',
-          marginBottom: 24,
-          width: '100%',
-          fontSize: '20px',
-          textAlign: 'center',
-          padding: 6,
-        },
-      }, [
-        date,
-      ]),
-
       h('style', `
 
         .markdown {
@@ -58,60 +32,94 @@ Notice.prototype.render = function () {
         }
 
         .markdown h1, .markdown h2, .markdown h3 {
-          margin: 10px 0;
-          font-weight: bold;
+          margin: 20px 0;
+          line-height: 18px;
+          font-weight: normal;
         }
 
         .markdown strong {
-          font-weight: bold;
+          font-weight: normal;
         }
         .markdown em {
           font-style: italic;
         }
 
         .markdown p {
-          margin: 10px 0;
+          margin: 20px 0;
+          line-height: 18px;
+          font-weight: normal;
         }
 
         .markdown a {
-          color: #df6b0e;
+          color: #8fdc97;
         }
 
-      `),
+        .markdown::-webkit-scrollbar {
+            width: 16px;
+        }
+         
+        .markdown::-webkit-scrollbar-track {
+            background-color: transparent;
+        }
+         
+        .markdown::-webkit-scrollbar-thumb {
+          background-color: #411a6a;
+          border-radius: 2px;
+          border: 6px solid transparent;
+          background-clip: content-box;
+        }
 
-      h('div.markdown', {
-        onScroll: (e) => {
-          var object = e.currentTarget
-          if (object.offsetHeight + object.scrollTop + 100 >= object.scrollHeight) {
-            this.setState({disclaimerDisabled: false})
-          }
-        },
-        style: {
-          background: 'rgb(235, 235, 235)',
-          height: '310px',
-          padding: '6px',
-          width: '90%',
-          overflowY: 'scroll',
-          scroll: 'auto',
-        },
-      }, [
-        h(ReactMarkdown, {
-          className: 'notice-box',
-          source: body,
-          skipHtml: true,
-        }),
-      ]),
+      `), [
+        h('h3.flex-center.terms-header', {
+          key: 'notice-key',
+          style: {
+            color: '#ffffff',
+            width: '100%',
+            fontSize: '16px',
+            textAlign: 'center',
+            margin: '20px 0px',
+          },
+        }, [
+          title,
+        ]),
+        h('div.markdown', {
+          key: 'notice-div-key',
+          onScroll: (e) => {
+            var object = e.currentTarget
+            if (object.offsetHeight + object.scrollTop + 100 >= object.scrollHeight) {
+              this.setState({disclaimerDisabled: false})
+            }
+          },
+          style: {
+            background: '#542289',
+            color: '#ffffff',
+            height: '310px',
+            width: '90%',
+            overflowY: 'scroll',
+            scroll: 'auto',
+            borderRadius: '3px',
+            fontSize: '14px',
+          },
+        }, [
+          h(ReactMarkdown, {
+            className: 'notice-box',
+            source: body,
+            skipHtml: true,
+          }),
+        ]),
 
-      h('button', {
-        disabled,
-        onClick: () => {
-          this.setState({disclaimerDisabled: false})
-          onConfirm()
-        },
-        style: {
-          marginTop: '18px',
-        },
-      }, 'Accept'),
+        h('button', {
+          key: 'notice-button-key',
+          disabled,
+          onClick: () => {
+            this.setState({disclaimerDisabled: true})
+            onConfirm()
+          },
+          style: {
+            marginTop: '18px',
+          },
+        }, 'Accept'),
+      ],
     ])
   )
 }
@@ -120,7 +128,7 @@ Notice.prototype.componentDidMount = function () {
   // eslint-disable-next-line react/no-find-dom-node
   var node = findDOMNode(this)
   linker.setupListener(node)
-  if (document.getElementsByClassName('notice-box')[0].clientHeight < 310) {
+  if (document.getElementsByClassName('notice-box')[0].clientHeight < 300) {
     this.setState({disclaimerDisabled: false})
   }
 }