+++ old-ui/lib/icon-factory.js
@@ -2,6 +2,7 @@ var iconFactory
 const isValidAddress = require('ethereumjs-util').isValidAddress
 const toChecksumAddress = require('ethereumjs-util').toChecksumAddress
 const contractMap = require('eth-contract-metadata')
+const colors = require('../../colors')
 
 module.exports = function (jazzicon) {
   if (!iconFactory) {
@@ -11,6 +12,7 @@ module.exports = function (jazzicon) {
 }
 
 function IconFactory (jazzicon) {
+  jazzicon.setColorsPalette(colors)
   this.jazzicon = jazzicon
   this.cache = {}
 }
@@ -37,7 +39,7 @@ IconFactory.prototype.generateIdenticonSvg = function (address, diameter) {
 // creates a new identicon
 IconFactory.prototype.generateNewIdenticon = function (address, diameter) {
   var numericRepresentation = jsNumberForAddress(address)
-  var identicon = this.jazzicon(diameter, numericRepresentation)
+  var identicon = this.jazzicon.generateIdenticon(diameter, numericRepresentation)
   return identicon
 }